app/requirements.txt
numpy==1.26.4
streamlit==1.32.1
pillow==10.2.0